Added simple try/catch/finally expression to CS parser and interpreter
The expression is of the form `try a catch b [finally c]`. If a `finally` clause is not given, `a` is evaluated and returned unless an error occurs, in which case `b` is evaluated and returned. If a `finally` clause is included, `a` is evaluated unless an error occurs, in which case `b` is evaluated; then `c` is evaluated and returned.